Connected the remote azure DB

+++ Client/ViewModels/IPostViewModel.cs
@@ -12,14 +12,14 @@ namespace BlazorCms.ViewModels
 {
     public interface IPostViewModel
     {
-        public long PostId { get; set; }
+        public int PostId { get; set; }
         public string PostTitle { get; set; }
         public string PostPermalink { get; set; }
         public string PostContent { get; set; }
         public string PostThumbnail { get; set; }
         public string PostCreated { get; set; }
         public string PostUpdated { get; set; }
-        public long PostAuthor { get; set; }
+        public int PostAuthor { get; set; }
         public string PostAuthorName { get; set; }
         public string Message { get; set; } 
         public string Display { get; set; } 
@@ -33,7 +33,7 @@ namespace BlazorCms.ViewModels
 
         public Task Update();
 
-        public Task Remove(long Id);
+        public Task Remove(int Id);
         public void OnImageRemovedSuccess(SuccessEventArgs args);
         public void OnImageUploadedSuccess(SuccessEventArgs args);
