+++ Server/CQSR/Commands/UpdatePostCommand.cs
@@ -6,12 +6,12 @@ namespace BlazorCms.Server.CQRS.Commands
 {
     public class UpdatePostCommand : IRequest<Post>
     {
-        public long PostId { get; set; }
+        public int PostId { get; set; }
         public string PostTitle { get; set; }
         public string PostPermalink { get; set; }
         public string PostContent { get; set; }
         public string PostThumbnail { get; set; }
-        public long PostAuthor { get; set; }
+        public int PostAuthor { get; set; }
         public string PostCreated { get; set; }
         public string PostUpdated { get; set; }
